# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sakura Beauty E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minRating
          in: query
          schema:
            type: number
        - name: sortBy
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /reviews/{productId}/eligibility:
    get:
      operationId: getReviewEligibility
      tags: [reviews]
      summary: Check if current user can review this product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Eligibility result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewEligibility"
  /reviews/{productId}:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: Get reviews for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review (auth required)
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{reviewId}:
    put:
      operationId: updateReview
      tags: [reviews]
      summary: Update own review
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "200":
          description: Updated review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{productId}/{reviewId}:
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review (admin or owner)
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        price:
          type: number
        discountPrice:
          type: ["number", "null"]
        category:
          type: string
        stock:
          type: integer
        description:
          type: string
        ingredients:
          type: ["string", "null"]
        keyBenefits:
          type: array
          items:
            type: string
        mainIngredients:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              icon:
                type: string
            required: [name, icon]
        bestFor:
          type: array
          items:
            type: string
        texture:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        averageRating:
          type: number
        reviewCount:
          type: integer
        isFeatured:
          type: boolean
        homepageSection:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - slug
        - price
        - discountPrice
        - category
        - stock
        - description
        - ingredients
        - keyBenefits
        - mainIngredients
        - bestFor
        - texture
        - images
        - averageRating
        - reviewCount
        - isFeatured
        - homepageSection
        - createdAt
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - totalPages
    MainIngredient:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
      required: [name, icon]
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        discountPrice:
          type: ["number", "null"]
        category:
          type: string
        stock:
          type: integer
        description:
          type: string
        homepageSection:
          type: ["string", "null"]
        ingredients:
          type: ["string", "null"]
        keyBenefits:
          type: array
          items:
            type: string
        mainIngredients:
          type: array
          items:
            $ref: "#/components/schemas/MainIngredient"
        bestFor:
          type: array
          items:
            type: string
        texture:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
      required:
        - name
        - price
        - category
        - stock
        - description
        - images
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        discountPrice:
          type: ["number", "null"]
        category:
          type: string
        stock:
          type: integer
        description:
          type: string
        ingredients:
          type: ["string", "null"]
        keyBenefits:
          type: array
          items:
            type: string
        mainIngredients:
          type: array
          items:
            $ref: "#/components/schemas/MainIngredient"
        bestFor:
          type: array
          items:
            type: string
        texture:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
...
    ReviewEligibility:
      type: object
      properties:
        canReview:
          type: boolean
        reason:
          type: string
          nullable: true
          enum: [already_reviewed, not_purchased, null]
      required:
        - canReview
    Review:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        userId:
          type: string
        userName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
      required:
        - id
        - productId
        - userId
        - userName
        - rating
        - comment
        - createdAt
    CreateReviewBody:
      type: object
      properties:
        rating:
          type: integer
        comment:
          type: string
      required:
        - rating
        - comment
...